Create basicTemplate.html
In this example, the ul element within the nav element is styled with list-style: none to remove the bullet points that normally appear with an unordered list. The li elements within the ul are styled with display: inline-block to make them appear inline with each other. The a elements within the li elements are styled with display: block to make them take up the full width of their parent li elements.

The margin-right property is used on the li elements to add space between each navigation item. This is optional and can be adjusted to your preference.

Note that there are many ways to align navigation elements horizontally in HTML and CSS, and the above example is just one of many possible solutions.